location screen in progress

+++ app/src/main/java/data/VehicleLocationData.kt
@@ -2,7 +2,7 @@ package data
 
 import com.google.gson.annotations.SerializedName
 
-data class VehicleLocationData(@SerializedName("Longitude") val longitude: Float?,
-                               @SerializedName("Latitude") val latitude: Float?,
-                               @SerializedName("Distance") val distance: Float?,
+data class VehicleLocationData(@SerializedName("Longitude") val longitude: Double?,
+                               @SerializedName("Latitude") val latitude: Double?,
+                               @SerializedName("Distance") val distance: Double?,
                                @SerializedName("timestamp") val timestamp: String?)
